few optimizations
- make sure docker image stays up-to-date
- use default values for platform and release
- added support for creating a patch from a git repo

+++ releases/createPatch.sh
@@ -6,23 +6,25 @@ NAME=
 ORIGINAL=
 PATCHED_DIR=
 DESCRIPTION=
+USE_GIT=no
 
 ###################### check for parameters and if they're valid ######################
 
 print_usage_and_exit() {
     echo
     echo "Creates a new patch by comparing two directories and storing the diff for a given release."
-    echo "Usage: $0 --release=<RELEASE> --original=<ORIGINAL> --patched=<PATCHED> [--name=<NAME>] [--desc=<DESCRIPTION>]"
+    echo "Usage: $0 --release=<RELEASE> --patched=<PATCHED> [--original=<ORIGINAL>] [--name=<NAME>] [--desc=<DESCRIPTION>]"
     echo "--release: the release version this patch applies to (e.g. 3.2.1)"
-    echo "--original: path to the original unmodified betaflight source, either a tarball or root directory (e.g. ../v3.2.1.tar.gz or betaflight-3.2.1)"
     echo "--patched: path to the modified source, must be a betaflight root directory (e.g. betaflight-3.2.1-patched)"
+    echo "--original: path to the original unmodified betaflight source, either a tarball or root directory (e.g. ../v3.2.1.tar.gz or betaflight-3.2.1)"
+    echo "            only necessary if the patched source is not a git repo"
     echo "--name: the patch name without the trailing '.patch' (optional, e.g. disablePidProfile)"
     echo "--desc: patch description (optional, e.g. \"Disable PID profile change via RC control\")"
     exit 1
 }
 
 clean_up() {
-    if [ -f $ORIGINAL ]; then
+    if [ -f $ORIGINAL ] && [ "$USE_GIT" != "yes" ]; then
         echo "Deleting $ORIGINAL_DIR..."
         rm -rf $ORIGINAL_DIR
     fi
@@ -60,27 +62,41 @@ if [ -z "$RELEASE" ]; then
 fi
 mkdir -p $RELEASE
 
-if [ -z "$ORIGINAL" ]; then
-    echo "No original path given with --original."
+if [ -z "$PATCHED_DIR" ]; then
+    echo "No patched path given with --patched."
     print_usage_and_exit
-elif [ ! -e $ORIGINAL ]; then
-    echo "The file/directory $ORIGINAL does not exist."
+elif [ ! -d $PATCHED_DIR ]; then
+    echo "The directory $PATCHED_DIR does not exist."
     exit 1
-elif [ -f $ORIGINAL ]; then
-    if [ -z "$(file $ORIGINAL | grep "compressed data")" ]; then
-        echo "The file $ORIGINAL does not seem to be a tarball."
+fi
+if [ -d "$PATCHED_DIR/.git" ]; then
+    USE_GIT=yes
+fi
+
+if [ "$USE_GIT" != "yes" ]; then
+    # this is not a git repo which means we need the original unmodified source
+    if [ -z "$ORIGINAL" ]; then
+        echo "No original path given with --original."
+        print_usage_and_exit
+    elif [ ! -e $ORIGINAL ]; then
+        echo "The file/directory $ORIGINAL does not exist."
+        exit 1
+    elif [ -f $ORIGINAL ]; then
+        if [ -z "$(file $ORIGINAL | grep "compressed data")" ]; then
+            echo "The file $ORIGINAL does not seem to be a tarball."
+            exit 1
+        fi
+        ORIGINAL_DIR=$(tar -tzf $ORIGINAL | head -1 | cut -f1 -d"/")
+        if [ -e $ORIGINAL_DIR ]; then
+            echo "Cannot untar $ORIGINAL. The directory $ORIGINAL_DIR already exists. Please rename this directory."
+            exit 1
+        fi
+    elif [ -d $ORIGINAL ]; then
+        ORIGINAL_DIR=$ORIGINAL
+    else
+        echo "$ORIGINAL is neither a tarball nor a directory."
         exit 1
     fi
-    ORIGINAL_DIR=$(tar -tzf $ORIGINAL | head -1 | cut -f1 -d"/")
-    if [ -e $ORIGINAL_DIR ]; then
-        echo "Cannot untar $ORIGINAL. The directory $ORIGINAL_DIR already exists. Please rename this directory."
-        exit 1
-    fi
-elif [ -d $ORIGINAL ]; then
-    ORIGINAL_DIR=$ORIGINAL
-else
-    echo "$ORIGINAL is neither a tarball nor a directory."
-    exit 1
 fi
 
 if [ -z "$NAME" ]; then
@@ -99,14 +115,6 @@ if [ -e $PATCH ]; then
     exit 1
 fi
 
-if [ -z "$PATCHED_DIR" ]; then
-    echo "No patched path given with --patched."
-    print_usage_and_exit
-elif [ ! -d $PATCHED_DIR ]; then
-    echo "The directory $PATCHED_DIR does not exist."
-    exit 1
-fi
-
 if [ -z "$DESCRIPTION" ]; then
     echo -n "Patch description: "
     read DESCRIPTION
@@ -118,7 +126,7 @@ fi
 
 ###################### compare and find differences between original and patched ######################
 
-if [ -f $ORIGINAL ]; then
+if [ -f $ORIGINAL ] && [ "$USE_GIT" != "yes" ]; then
     echo "Extracting $ORIGINAL..."
     tar xf $ORIGINAL
 fi
@@ -127,7 +135,11 @@ echo "Creating patch $PATCH..."
 echo "$DESCRIPTION" > $PATCH
 echo >> $PATCH
 set +e
-diff -Naur $ORIGINAL_DIR $PATCHED_DIR >> $PATCH
+if [ "$USE_GIT" != "yes" ]; then
+    diff -Naur $ORIGINAL_DIR $PATCHED_DIR >> $PATCH
+else 
+    git -C $PATCHED_DIR diff --exit-code >> $PATCH
+fi
 if [ $? = 0 ]; then
     echo "No differences!"
     rm -f $PATCH
